fixing the error
