+++ src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
 import userReducer from "../features/user/userSlice";
-import movieReducer from "../features/user/movie/movieSlice";
+import movieReducer from "../features/movie/movieSlice";
 
 export default configureStore({
     reducer: {
