+++ src/components/Home.js
@@ -7,14 +7,13 @@ import Originals from './Originals';
 import Trending from './Trending';
 import { useEffect } from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-
 import db from '../firebase';
-import { setMovies } from '../features/user/movie/movieSlice';
+import { setMovies } from '../features/movie/movieSlice';
 import { selectUserName } from '../features/user/userSlice';
 
 const Home = (props) => {
   const dispatch = useDispatch()
-  const useName = useSelector(selectUserName);
+  const userName = useSelector(selectUserName);
   let recommends = [];
   let newDisneys = [];
   let originals = [];
